requirements.txt
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
cloudpickle==2.2.1
Command==0.1.0
filelock==3.12.2
google-auth==2.19.1
gym==0.26.2
gym-notices==0.0.8
idna==3.4
importlib-metadata==6.8.0
Jinja2==3.1.2
kubernetes==26.1.0
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
oauthlib==3.2.2
Pillow==10.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sympy==1.12
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
typing_extensions==4.7.1
urllib3==1.26.16
websocket-client==1.5.3
zipp==3.16.2
